created new jira admin middleware and tests to return 403 when no auth
