+++ src/middleware/jira-symmetric-jwt-middleware.test.ts
@@ -3,9 +3,8 @@ import express, { Application, NextFunction, Request, Response } from "express";
 import { noop } from "lodash";
 import supertest from "supertest";
 import { getLogger } from "~/src/config/logger";
-import { jiraSymmetricJwtMiddleware, setJiraAdminPrivileges } from "~/src/middleware/jira-symmetric-jwt-middleware";
+import { jiraSymmetricJwtMiddleware } from "~/src/middleware/jira-symmetric-jwt-middleware";
 import { Installation } from "~/src/models/installation";
-import { DatabaseStateCreator } from "test/utils/database-state-creator";
 
 jest.mock("config/feature-flags");
 const testSharedSecret = "test-secret";
@@ -227,73 +226,4 @@ describe("jiraSymmetricJwtMiddleware", () => {
 		return app;
 	};
 
-});
-
-describe("setJiraAdminPrivileges",  () => {
-	const mockRequest = {
-		session: {},
-		log: {
-			info: jest.fn(),
-			debug: jest.fn(),
-			error: jest.fn()
-		}
-	} as unknown as Request;
-	const mockClaims = { sub: "1111" };
-	let installation;
-
-	beforeEach(async () => {
-		installation = (await new DatabaseStateCreator().create()).installation;
-	});
-
-	it("sets session isJiraAdmin to true if user has ADMINISTER permission", async () => {
-		mockRequest.session.isJiraAdmin = undefined;
-		const payload = {
-			accountId: "1111",
-			globalPermissions: [
-				"ADMINISTER"
-			]
-		};
-		jiraNock
-			.post("/rest/api/latest/permissions/check", payload)
-			.reply(200, { globalPermissions: ["ADMINISTER"] });
-
-		await setJiraAdminPrivileges(mockRequest, mockClaims, installation);
-
-		expect(mockRequest.session.isJiraAdmin).toBe(true);
-	});
-
-	it("sets session isJiraAdmin to false if user does not have ADMINISTER permission", async () => {
-		mockClaims.sub = "2222";
-		mockRequest.session.isJiraAdmin = undefined;
-		const payload = {
-			accountId: "2222",
-			globalPermissions: [
-				"ADMINISTER"
-			]
-		};
-		jiraNock
-			.post("/rest/api/latest/permissions/check", payload)
-			.reply(200, { globalPermissions: [] });
-
-		await setJiraAdminPrivileges(mockRequest, mockClaims, installation);
-
-		expect(mockRequest.session.isJiraAdmin).toBe(false);
-	});
-
-	it("should exit early when claim has no sub", async () => {
-		const mockClaimsNoSub = {};
-		mockRequest.session.isJiraAdmin = undefined;
-
-		await setJiraAdminPrivileges(mockRequest, mockClaimsNoSub, installation);
-
-		expect(mockRequest.session.isJiraAdmin).toBe(undefined);
-	});
-
-	it("should return session value without JiraClient request if already exists", async () => {
-		mockRequest.session.isJiraAdmin = "true";
-
-		await setJiraAdminPrivileges(mockRequest, mockClaims, installation);
-
-		expect(mockRequest.session.isJiraAdmin).toBe("true");
-	});
 });
