+++ src/middleware/jira-symmetric-jwt-middleware.ts
@@ -5,32 +5,7 @@ import { getJWTRequest, TokenType, validateQsh } from "~/src/jira/util/jwt";
 import { Installation } from "~/src/models/installation";
 import { moduleUrls } from "~/src/routes/jira/atlassian-connect/jira-atlassian-connect-get";
 import { matchRouteWithPattern } from "~/src/util/match-route-with-pattern";
-import { JiraClient } from "models/jira-client";
-
-export const setJiraAdminPrivileges = async (req: Request, claims: Record<any, any>, installation: Installation) => {
-	const ADMIN_PERMISSION = "ADMINISTER";
-	// We only need to add this to the session if it doesn't exist
-	if (req.session.isJiraAdmin !== undefined) {
-		return;
-	}
-
-	try {
-		const userAccountId = claims.sub;
-		// Can't check permissions without userAccountId
-		if (!userAccountId) {
-			return;
-		}
-		const jiraClient = await JiraClient.getNewClient(installation, req.log);
-		// Make jira call to permissions with userAccountId.
-		const permissions = await jiraClient.checkAdminPermissions(userAccountId);
-		const hasAdminPermissions = permissions.data.globalPermissions.includes(ADMIN_PERMISSION);
-
-		req.session.isJiraAdmin = hasAdminPermissions;
-		req.log.info({ isAdmin :req.session.isJiraAdmin }, "Admin permissions set");
-	} catch (err) {
-		req.log.error({ err }, "Failed to fetch Jira Admin rights");
-	}
-};
+import { setJiraAdminPrivileges } from "middleware/jira-admin-permission-middleware";
 
 export const jiraSymmetricJwtMiddleware = async (req: Request, res: Response, next: NextFunction) => {
